feat: 알림 관리를 위한 useNotification 훅 추가
