+++ src/basic/features/show-notification/index.ts
@@ -1,0 +1,1 @@
+export { useNotification } from "./hooks/useNotification";
